@@ -2,26 +2,26 @@
 name = "swiftnav"
 version = "0.10.0"
 authors = ["Swift Navigation <dev@swiftnav.com>"]
-edition = "2018"
+edition = "2024"
 description = "GNSS positioning and related utilities"
 readme = "README.md"
 repository = "https://github.com/swift-nav/swiftnav-rs"
 license = "LGPL-3.0"
-rust-version = "1.85.0"
+rust-version = "1.87.0"
 
 [dependencies]
-rustversion = "1.0"
-chrono = { version = "0.4", optional = true }
-strum = { version = "0.27", features = ["derive"] }
-nalgebra = "0.33"
-thiserror = "2.0"
+bon = "3.8.1"
+rustversion = "1.0.22"
+chrono = { version = "0.4.42" }
+strum = { version = "0.27.2", features = ["derive"] }
+nalgebra = "0.34.1"
+thiserror = "2.0.17"
 
 [dev-dependencies]
 float_eq = "1.0.1"
-proptest = "1.5"
+proptest = "1.9.0"
 
 # This tells docs.rs to include the katex header for math formatting
 # To do this locally 
 [package.metadata.docs.rs]
-rustdoc-args = [ "--html-in-header", "katex-header.html" ]
-
+rustdoc-args = ["--html-in-header", "katex-header.html"]
