@@ -8,7 +8,7 @@
 // EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED
 // WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.
 
-use super::{consts, Constellation};
+use super::{Constellation, consts};
 
 /// Code identifiers
 #[derive(
